multiple refs can now use the same server
Given the simplicity of the game rules, this is a situation where
the server can handle many games simultaneously. The ref now knows
how to extract the server game number from player 0's output and
tell player 1 the proper game to connect to. Slowdowns are expected
to be the fault of the client.