server -arena command line flag shuts off sending gamelogs to clients
The gamelogs are pretty huge, and when the arena is running full tilt,
the server may be asked to send 20 of them within a few seconds, to disappear
into the ether. So. Rather than have to explain to the client that it
shouldn't ask for a gamelog, we're just going to send it a blank.