Big Score Board update, Yay! Now we have each team's virus drawn next to their name, and (being much more useful) we have a bar representing how many bytedollars the two teams have next to each other so you can get a quick and easy pictorial representation of thier progress relative to one anouther. This is similar to the one in MegaMiner 6: Modular FYI.
In the future I'd like to draw a texture on the bars and make the top score
board look better, but it does it's job well enough right now.

Oh, I also updated my AI. It's not half bad now (IMHO); it's still a pretty
slow when it starts to rape though.