Tournament Scheduler. First Draft.
Added doodads to various models to provide things the tournament
scheduler needs. Added a Match model, representing a first-to-N series.
And of course, a tournament scheduler. Right now I'm building up the
tournament brackets by hand. This is tedious. Also, the scheduler is
doing a depth first search through the bracket, which ain't quite right.